Remove sprockets from gemspec.
We already depend on actionpack, which itself depends on sprockets.

Having sprockets in the gemspec, tied to a specific version, creates the
potential for version conflicts with the version of sprockets that
actionpack wants to load.

For example, we want to bump sprockets to 2.1.0 for the Rails 3.1.2
release, but this cannot be done with the existing sass-rails releases,
because they depend on sprockets ~> 2.0.0.