Allowing railtie and compressor to honor config settings for sass.style and assets.css_compressor with default values if not otherwise set.  This allows assets.css_compressor = :yui to actually work!
Setting ||= :compressed is pointless if the style was initialized to :expanded at the beginning of the rail ties file.  So it had to be moved to the else clause of the setup_compression initializer.

Added an initializer to CssCompressor to allow for any compression style to be passed in, and then passing in from the railtie.  It might be good to do the same for the other options in the compressor.  The initialize method is already heading in this direction.

Added another test fixture app, with different config settings set in deferent environments.  Calling runcmd with different environments then allows for different config paths to be run.