Fix errors from gem defaults
Gems are no longer allowed to leave the default TODO's in place. This
replaces all the ones that can't be TODO anymore with something that is
acceptable.

```
Error:
bundler: failed to load command: rails
(Users/eileen/.rbenv/versions/2.3.0/lib/ruby/gems/2.3.0/bin/rails)
Gem::InvalidSpecificationException: The gemspec at
/private/var/folders/fy/rml6_y9x3b392jpmpy160g7r0000gn/T/d20160630-25961-h3mf6o/engine_project.gemspec
is not valid. Please fix this gemspec.
The validation error was '"FIXME" or "TODO" is not an author'
```