Prevent beta versions of gems as a dependency
Before, the version dependency for example on sprockets was like this:

    s.add_dependency 'sprockets',       '>= 2.8', '< 4.0'

But this allowed to install beta versions of sprockets like current
`4.0.0.beta3` (because `4.0.0.beta3 < 4.0). Which is not compatible with
sass-rails.

By using `< 4.x`, it does sucessfully prevent to install any version 4
of sprockets, including beta versions.

Other gem dependencies changed accordingly.